requirements.txt
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
dataclass-wizard==0.22.2
distlib==0.3.6
filelock==3.12.0
flake8==6.0.0
identify==2.5.24
idna==3.4
iniconfig==2.0.0
isort==5.12.0
Levenshtein==0.21.0
mccabe==0.7.0
munch==3.0.0
munch-stubs==0.1.2
mypy==1.3.0
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.2
pandas-stubs==2.0.2.230605
pathspec==0.11.1
platformdirs==3.5.1
pluggy==1.0.0
pre-commit==3.3.2
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
pyodc==1.3.0
pytest==7.3.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
rapidfuzz==3.1.1
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.23.1
ruff==0.0.271
six==1.16.0
tqdm==4.65.0
types-oauthlib==3.2.0.8
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.10
types-requests==2.31.0.1
types-urllib3==1.26.25.13
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.23.0
watchdog==3.0.0
